Add module structure
